+++ NEWS.md
@@ -2,7 +2,7 @@
 
 ## New features
 
-* New movie: mean_vs_median.  Compares the sampling distributions of the mean and median for random samples from either a standard normal distribution or a standard Student t distribution.
+* New movie: `mean_vs_median()`.  Compares the sampling distributions of the mean and median for random samples from either a standard normal distribution or a standard Student t distribution.
 
 ## Bug fixes and minor improvements