+++ NEWS.md
@@ -1,4 +1,4 @@
-# smovie 1.0.1.9000
+# smovie 1.1.1
 
 ## New features